# src/docs/swagger.yaml
definitions:
...
  main.userModifyErrorResponse:
    properties:
      message:
        type: string
      segments_add:
        items:
          type: string
        type: array
      segments_remove:
        items:
          type: string
        type: array
    type: object
  main.userSegmentsModifyBody:
    properties:
      segments_add:
        items:
          type: string
        type: array
      segments_remove:
        items:
          type: string
        type: array
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /modify-user-segments:
    put:
      description: |-
        user and segments exist in the database.
        if not, error is returned.
      parameters:
      - description: Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.userSegmentsModifyBody'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.userModifyErrorResponse'
      summary: modifies users segments
  /user-report/{filename}:
    get:
      parameters:
      - description: Filename
        in: path
        name: filename
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: downloads report ebout user activity
swagger: "2.0"